// circabc-frontend/src/app/core/generated/circabc/model/eventItemDefinition.ts
/**
 * CIRCABC REST API
 * This is the first version of the CIRCABC REST API used by the new User Interface 
 *
 * The version of the OpenAPI document: 4.2.0
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * definition of an event or meeting. Used to retrieve information about it
 */
export interface EventItemDefinition { 
    id?: string;
    appointmentContainerId?: string;
    igId?: string;
    appointmentType?: EventItemDefinition.AppointmentTypeEnum;
    title?: string;
    appointmentDate?: string;
    startTime?: string;
    endTime?: string;
    contact?: string;
    meetingStatus?: string;
    eventAbstract?: string;
    location?: string;
    occurrenceRate?: string;
    timeZone?: string;
}
export namespace EventItemDefinition {
    export type AppointmentTypeEnum = 'Event' | 'Meeting';
    export const AppointmentTypeEnum = {
        Event: 'Event' as AppointmentTypeEnum,
        Meeting: 'Meeting' as AppointmentTypeEnum
    };
}